added a utils file
